Extract HNSW code
